+++ examples/tcp_server.c
@@ -21,7 +21,7 @@ int main(int argc, char *argv[]) {
 		return 1;
 	}
 
-	tfi_server *server = tfi_start_server(atoi(argv[1]), TFI_TCP, 1);
+	tfi_server *server = tfi_start_server((short)atoi(argv[1]), TFI_TCP, 1);
 	tfi_client *client = tfi_accept(server);
 	tfi_client_thread *ct = tfi_start_client_thread(client, connection_func);
